CHORE: Rename lib as utils
